completely revamp imap engine and start imap_ditm
IMAP engine blindly writes imap cmds/resps that its given and blindly
passes out imap cmds/resps that it reads off the wire.

Remove the manual multithreading in the IMAP engine which previously was
meant to support the actual application logic.

In support of the new passthru imap strategy rather than the subserver
imap strategy, halt work on sm_fetch, which will no longer offer
meaningful code reuse for the real imap ditm.

Instead, begin work on the real imap_ditm.  Configure the multithreading
logic to be based around using the uv_loop directly, and enqueuing work
directly on the uv_loop with normal uv_work_t requests.

TODO: investigate moving the tls_engine_t and the imap engine to just
operating on the uv_loop thread.  Probably some profiling effort is
necessary before an informed decision can be made as to wether or not
the complexity/performance tradeoff is worth it.