abandon imap_ditm and passthru strategy, rework sm_fetch
Passthru IMAP DITM quickly became untenable.  Revert to the original
subserver strategy, except take the subserver startegy further; maintain
local modseq numbers and everything (not local UIDs yet).

Presently the imap maildir owns the per-mailbox synchronization logic;
sm_fetch owns the imap session upwards, but it passes control of the
session to the imaldir_t, and imaildir_t automatically does SELECT and
starts synchronizing the mailbox.  When sm_fetch sees the "synchronized"
signal, it requests the imaildir_t release the session, and imaildir_t
does CLOSE and returns it.

In order to keep state of IMAP message metadata, this change introduces
an LMDB dependency.  LMDB is only used for an easy, persistent key-value
store; it would be easy to remove LMDB for something homespun if LMDB
proves to be a poor choice.

Presently, imaildir_t is intended to be safely multi-threaded, but in
reality its threading model needs some serious reworking.