rethought deletion state and fix vanished response
There's still some memory errors to track down but sm_fetch is almost
fully functional as a one-shot imap downloader/resyncer.  In its current
form it can identify the difference between a missing file that hasn't
been downloaded and a missing file that was downloaded by the user, and
an expunged message that was pushed to the server vs one that didn't get
pushed to the server.

On the download side, QRESYNC seems to be working very nicely.