add sm_serve and reconfigure imape_t's downstream
In the case of sm_serve, imape_t needs to send a quit message to a
single entity, (a manager that tracks the lifetime of all server_t
instances), but each imape_data_t's downstream is just the server_t
instance, so those are now distict.

TODO: Most of the complexity of the QUIT sequence can be done away with
in favor of a single engine manager that doesn't let any engine threads
actually exit until all the other engines have reported that they are
dead.