refactor keypair_t to be a shared pointer
keypair_t is now a shared pointer to immutable memory which is
automatically freed when the last keypair_t is freed.  The keypair_t
itself uses link_t to support lists of keys very easily.

Revamp the encrypter_t API accordingly.  It is much easier to use now.

Refactor encrypt_msg.c accordingly.

Introduce test/mariadb.py which has a context manager for running
mariadb as the local user and under a local directory.  This was
necessary to be able to test the changes to encrypt_msg.c.

Write end-to-end tests for encrypt_msg and decrypt_msg, where
encrypt_msg can even be tested against the mariadb.py server.

Bring pysm.c into this repository so that it could be used as part of
the enc_dec.py end-to-end test.

Fix the use of global compiler/linker options in cmake so that pysm.so
could be built without asan while the rest of the project has asan.