fix bootstrapping behavior of up_t
up_t had some complex and likely implicit handling of inital
synchronizations (I think it didn't handle the UIDVALIDITY changing case
properly, only the initial synchronization case).  In any case, it was
confusing and it was definitely not going to work nicely with the
upcoming download pause behavior to support APPEND and COPY.

This change makes the bootstrapping behavior both simpler and more
explicit, by using a single UID FETCH ... (CHANGEDSINCE 0 VANISHED)
to bootstrap the mailbox rather than a combination of UID SEARCH and
UID FETCH (which did not previously populate expunges anyway).