support download holds in libimaildir
Pausing downloads at a mailbox name level is required for APPEND and
COPY to work without redownloading what we submit to the server.

The dirmgr coordinates the holds for all mailboxes, since the holds must
be stored independently of the lifetimes of the imaildirs.  That part is
good.

Then, there must be support for adding messages to a mailbox, whether or
not that mailbox is open.  That part is pretty awfully complicated and
I really dislike it, because you have to support adding files whether or
not the relevant mailbox is open.  imaildir_t now supports opening in
"lite" mode which allows you to add files and nothing else (gross).
Additionally, the mechanisms for adding a file are slight different
(read: a whole new function) than the old mechanisms, which were based
on receiving successive FETCH commands.

But I haven't been able to come up with anything better yet.