fix literal handling
Reintroduce LITERAL_END dummy token so that RAW-after-literal is
properly rejected.

Introduce IMAP_CMD_PLUS and IMAP_RESP_PLUS types which allow easy
in-order processing of literal synchronizations from clients.

Write all literals with the LITERAL+ notation.

Also, introduce the -Wswitch-enum and fix a couple of bugs that were
discovered in the process.