fix e2e_citm.py to prevent hangs
ReaderThreads and SocketReaderThreads both had to be started after the
subprocess was started, but the subprocess had to be killed in order to
reliably close thier blocking reads.  Using select or poll was not an
option because that would never have worked on windows (the ReaderThread
is not reading a socket) so the ReaderThread and SocketReaderThread now
take a closable object that they call close() on at exit if they are
shutting down with a failure.