support APPEND commands
APPEND commands are received by the server_t, then intercepted by the
sf_pair, where the unencrypted copy is stored to local disk.  The
sf_pair requests a hold from dirmgr.  Then the sf_pair modifies the
APPEND to be encrypted and relays it through the fetcher_t.

Also add some basic tests for the APPEND command.

This APPEND functionality has an obnoxious bug though: the MODSEQ of the
message is wrong when appending to an imaildir that is not fully
up-to-date with the remote mailserver.  Even if you are fully
up-to-date, that imaildir would still need to check that the UID for the
message was a valid next modification to the box, as another
simultaneous APPEND could be in flight from some other source.  There
are two approaches to solving this that I see:

 1. Keep the content of the APPENDed message around somewhere and use it
    to skip the download for that UID at the time that you would
    normally download that UID.

    This seems tricky and would add yet another layer of state to the
    process of adding messages to a mailbox.  In particular, this layer
    of state would have to persist between process invocations, so it
    would be extra bad.  It also does not fit anywhere in the proper
    maildir specification for how to store mail in a mailbox, so it
    wouldn't play well with e.g. mutt.  I don't like this option.

 2. Keep a local ordering of UIDs, with a mapping of local UIDs to
    remote UIDs.  This releives us of having to synchronize our local
    source of truth (modseq numbers) with the remote source of truth
    (uids).  I have thought I should probably do this for a while, so
    this is definitely the better option.