cleanup '+' handling
Previously when a '+' was needed, the parser would call a special
need_plus callback.  Because the '+' was a sync point, the need_plus
callback necessarily had to pass the request through the same queues as
normal imap commands to ensure proper synchronization.

Now we skip the need_plus() special callback nonsense and just call the
normal cmd() callback with an IMAP_CMD_RESP_PLUS-type command.