migmysql: like golang-migrations but better
migmysql embeds the downgrade scripts into the database, so it is always
possible to downgrade, even when the downgrade logic is no longer on the
filesystem.

This is important when you want to git-checkout old code and migrate the
database (as opposed to figuring out how to run the migrations manually
before the git checkout before doing the git-checkout).