fix out-of-order reselect bug
Reselects were sent as normal relays because it was known they had to be
well-ordered with respect to other relayed commands.  However, since the
imaildir would request a reselect that only got enqueued at a later
time, it was possible additional relayed commands got enqueued before
the reselect, even though the reselect was requested first.

This change fixes that problem without allowing multiple, pipelined
reselects (one of the original design goals of the reselect logic) by
having a secondary queue for relay requests that come in after a
reselect is requested.  That queue is appeneded to the normal relay
queue as soon as the reselect is queued.