work on removing bison
Bison generates GPL'd C code, which I don't want to commit into my
repository.  But I do want to commit the parser code into my repository
so that it's easy for new devs to build the code right away.

I tried lemon, but it has a sort of lazy execution model that makes it
hard for me to get the side-effects to execute at the right time.

I wanted to try antlr, but it stopped generating C code in v4.9.2.

I'm working on my own parser generator, a recursive-descent generator,
which seems promising.  Of course, if this worked, it's all in basic
python and it wouldn't require committing parser into the repository.