fix code generation for windows
Windows can't run bison at all, at least not any recent version of
bison.  We can generate code inside the source directory on linux (which
can be shared with windows via samba) so windows can mooch off of
linux's working bison install.

An easier and cleaner solution would be to just commit the generated
code into the repository.  That way windows users would be able to just
clone the repo and build the software (even if they couldn't reliably
develop certain files).  However, the code bison generates is GPL'd,
with a caveat that the code can be distributed in non-GPL projects so
long as those projects are not themselves parser generators making use
of the parser skeleton from bison to do the parser generation.
Splintermail obviously qualifies as not a parser generator at all.

Nonetheless, I'm not willing to commit even that GPL'd code into the
source code, as it seems unnecessarily risky and I'm not quite sure how
to document it properly.  Better to only ship the compiled version with
binaries.

A significant amount of work has been put into finding a suitable bison
replacement.  That is a long-term goal but it will be a ton of work and
shouldn't block the release of IMAP support in the near-term.