gen.py: rework fallbacks and add tests
Fallbacks can cause some conflicts which may or may not be resolved by
runtime preference.  Imagine:

    %fallback LETTER A

Now the behavior of matching a LETTER will change, possibly due to
non-local context.  That is, an expression starting with LETTER can
start with an A token, unless that expression is inside of a larger
branch statement, where a separate branch can start with an A
explicitly.  This is a fallback-caused first/first conflict.

Similarly, an expression starting with LETTER can start with an A token,
unless that expression follows an A* statement. Or worse, a sequence of
(LETTER* A) will never match the final A.  This is a fallback-caused
first/follow conflict.

In both of these cases, there is some tricky non-locally-obvious
behavior where expressions have different behavior based on how they are
called.  Additionally, if the tricky behavior were considered acceptable
to the user, the generator would still have to be able to properly
resolve each possible conflict.

As it happens, the first/first case is pretty easy to resolve in a
recursive descent parser.  In branching cases, you simply prefer the
more specific token meaning to the more general, and the information of
which tokens match the branches you are considering is known statically.

The first/follow case is pretty hard to resolve though; in order to
prefer the specific token meaning to the general, the generated parser
would have to know at runtime which possible tokens could follow the
function currently executing.  This might add a lot of complexity to the
generated parser (I for one have no plan to implement this).

Therefore, the grammar check now allows the first/first case silently,
but disallows the first/follow case.  Some day that should be
configurable so that users can pick and choose the strictness they care
about, possibly according to the capabilities of some external
generator.