libimaildir: refactor modseq_dn logic
After hitting a double-modseq bug causing a memory error, but being
completely unable to reproduce it, refactor the modseq logic to be
simpler and more robust.  Now the highest modseq down is tracked not by
examining current msg_mod_t's, but based on an always-increasing,
persisted value for the mailbox, much like the himodseq_up or the
hi_uid_up.

In the proccess, fix the long-standing issue with giving out "ghost"
modseq numbers for STOREs to already-expunged messages.  Before, there
was no way to track those ghost modseq numbers, but now there is.

Also found a few new cornercases which were not previously handled, but
are handled now, and refined/solidified the msgs/expunged/mods tree
structure and explained it in comments in imaildir.h.