libimaildir: prevent obscure race condition
There was a rare race condition where the up_t might receive an
unsolicited FETCH response for a UID that is being APPENDED by another
session.  Normally, the APPEND from another session we control would
only cause an EXISTS response for this up_t, but if a third-party
session adds a flag we could still see an unsolicited FETCH response.

If that happened and we weren't careful, we would create an unfilled
msg_t in up.c::fetch_resp(), meanwhile the APPENDing session would
create another msg_t unconditionally via imaildir_add_local_file() as it
processes the relayed APPEND response.

Since the APPEND response is passed around with asynchronous wakeups, it
is possible for an up_t to process a new response before the APPEND
response is fully processed.  Therefore the race condition is actually
possible (though it must be spectacularly rare).

The easiest solution was to leave up.c::fetch_resp() unmodified but
detect the race condition in imaildir_add_local_file().