libimaildir: one FETCH response per uid per msg
Before, when a FETCH command was issued, if a message that was requested
by the FETCH also had a flag update, that message would have two FETCH
responses, one with just the FLAG update and another with the requested
FETCH parameters.  Now, only one FETCH response is issued per message,
and it will include FLAGS if there were flag updates to report. This is
exactly how dovecot behaves in this case.

Also, this fixes a bug where new messages were added before the
canonical uids dn were added, so a client could actually FETCH messages
that shouldn't be in its view yet.