gen.py: fix bugs in the C generator error messages
Groupmask was not getting cleared in all the cases where the mask was
being cleared, which would have cause error messages to claim a larger
number of tokens would have been accepted than in reality.  This was a
bug in the recently added TokenCheckOptimizer.

Also, mask information was not being set properly when passing through
branch statements containing an %empty branch, causing error messages to
claim that fewer tokens would have been accepted than in reality.  This
was a bug that has existed since support for %empty branches was added.