stream_i: make reads and writes a symmetric API
Changes to stream_i interface:
  - read/read_cb instead of read_start/read_stop/alloc_cb/read_cb
  - return derr_t in await_cb
  - reads and writes use dstr_t instead of uv_buf_t

The change to read/read_cb simplifies usage and implementation of a
stream significantly; rather than the user calling read_start() and
read_stop(), then writing an alloc_cb() which interacts with a
read_cb(), just have the user call read() when they want to read, and
they get a read_cb() when it's done.

The duv_passthru_t translates the stream_i mechanics into libuv
mechanics.

A key part of duv_passthru_t being able to avoid extraneous read_stop()
calls is the introduction of a scheduler_i.  Also, the scheduler_i makes
it possible for most stream_i implementations to operate without libuv
at all.  There is a duv_scheduler_t that requires a uv_loop, but only
the duv_passthru_t should ever need it; other stream_i implementations
should work just fine with a generic scheduler_i.

Returning a derr_t return value in the await_cb simplifies the stream
API by removing the strerror and err_name methods, and improves overall
error handling, at the cost of memory allocation occuring during errors.

Using dstr_t in read and write calls makes so that the splintermail
codebase is more coherent.  Before, libduv was sorta halfway trying to
be independent of libdstr, but the separation wasn't that good anyway,
and combining the error handling and the string library into the
interface is much nicer on the splintermail codebase.