fix stream->await() mechanics
The previous strategy of providing an await_cb to the first layer of a
pipe, and having it be intercepted and then called from later pipe
layers did not work because later layers of the pipe would not know
which stream_i to pass in as an argument.

Instead, separate the await() call from the initialization.  A complex
pipeline or network of streams can be initialized all at once, and only
the "ends" which the application will interact with directly need to be
awaited.