introduce stream->wrapper_data
There are two general strategies for regaining the object who planned a
callback from the callback itself.

The first is CONTAINER_OF, and the second is a void *data parameter.

CONTAINER_OF works for the owner of a stream, but it does not work with
a stream wrapper, which does not own the memory of the stream_i.

However, currently the .data is owned by the owner of the stream.
Transferring ownership is not possible, because in a STARTTLS situation,
the owner may have some pending reads at the time that the tls wrapper
wraps the base stream.  The tls wrapper needs to set .data at
initialization time, when it awaits the base.  But then the pending
read_cbs might fire and the owner will find .data to be corrupted.

The solution is to have a void* on the stream_i which is dedicated to
the wrapper of a stream.  You could reserve .data for wrappers and
require owners to use CONTAINER_OF instead.  That's a bit inflexible.
Instead, we introduced .wrapper_data alongside .data.

Additionally, .data and .wrapper_data are right on the stream_i struct
rather than being a part of the interface, which cuts down on
boilerplate.