stream_i: convert close() into cancel()
await_cb is now called automatically after EOF and shutdown_cb with no
error.  stream->cancel() will cause await_cb to gives E_CANCELED.

This is more consistent with other asynchronous apis in the system,
which usually have apis like:

    do_thing(&thing, thing_done_cb)

and
    thing_cancel(&thing)

The difference is that the stream_i has an api that is active before the
await_cb.  But before, close() was required for both the preemptive case
and the success case (after a bidirectional shutdown).  This was extra
annoying becase the read() and write() apis were already broken by the
time the bidirectional shutdown completed.  But cancel() is only for the
preemptive case.