fix use-after-free bug
server_t, fetcher_t, and keysync_t each sent a close event_t, but none
of them had a reference associated with it, so in rare cases a close
event would queue work that got called after the object was freed.

The solution is to just start with one extra ref on each object, since
the close event is guaranteed to happen, and ref_dn after it's done.