logger.h: PROP never clears an error
PROP and MERGE both have error-merging behaviors.  Previously, PROP
always took new error (even if it was E_OK) while MERGE always preferred
the old error (but overrode it when it was E_OK).

This required more careful error handling when dealing with errors
coming in from await_cb's, because this innocent looking code:

    void await_cb(stream_i *s, derr_t e){
        thing_t *t = s->data;
        PROP_VAR_GO(&t->e, e, fail);
        ...
    }

would _erase_ t->e if e was E_OK.  Careful thinking and manual merging
was required to obtain a logical OR of errors.