fix detach semantics of limit and chunked rstreams
If the underlying is already awaited when the limit or chunked rstream
is ready to exit, the limit or chunked rstream has no way to tell.  It
can only tell if its immediate parent has been awaited, but there could
be a delay.

Since the duv_http_req_t is what is going to take ownership of the the
underlying stream->await() the instant that the limit or chunked rstream
wants to detach, the best option is to just have a try_detach() callback
that limit or chunked rstream can call, which returns true if it was
successful.

This is annoyingly intricate but at least the limit and chunked rstream
are internal to the libhttp module.