libhttp: reader automatically left-shifts buffer
The requirement that all headers fit in a fixed buffer is a pretty silly
requirement, so we got rid of it.

Also, after some careful though, the http_reader_t's scanner state was
sometimes inconsistent after encountering a wantmore=true token.  Now
the scanner state is reset in those cases so the bug is gone.

chunked_rstream_t also detects wantmore=true but doesn't preserve
scanner state at all, so it was unaffected by that bug.