libhttp: borrow has better eof and await behaviors
The borrow_rstream should never finish before its base is awaited,
though.  It also shouldn't be aware of detach mechanics.  If the end of
the pipe calls try_detach() successfully then the borrow should just be
abandoned without ever calling await_cb.

Also, when the borrow sees an eof, it should cancel the base since
borrow is only used for http and if the other end does a shutdown,
there's no point in trying to write another request later.