libhttp: chunked_rstream rejects eof and extra data
The chunked_rstream should never see an EOF, even in
"Connection: close" situations.

Instead, the last read it makes should return a complete, well-formed
chunk, and it should not call read again to see the EOF.

An EOF therefore now triggers an E_RESPONSE message, since the chunked
encoding was incomplete.

Also, the chunked_rstream should not ever read additional content after
the trailer is complete, because that would be the start of a new
response (without a corresponding request).  Such a check is not
necessary in the close-delineated message case, because the EOF is
always reached, and it is not necessary in the limit_rstream case,
because the limit_rstream only requests the exact right number of bytes.

This change also adds failure tests to test_chunked.c.

chunked