fix dense log file with large initial VANISHED
When a large VANISHED response existed during an initial sync, writing
the initial log file might take a nearly unbounded number of fsyncs.
There are other performance gains to be had with the log file, but by
far the worst offender was all the empty messages with no information
about them, so just get rid of them.

- The imaildir_t avoids creating entries for them, both in memory and in
  the log file.
- The logger ignores them when it encounters them in a legacy log file.
- The logger drops them the first time that it compresses the log file.