update server/acme/DESIGN.md
The original ACME design was for the splintermail server to own one
ACME account and create certificates on behalf of end-users.  However,
handing out certificates to end users would violate the terms of service
of any ACME provider, since that is creating a certificate but failing
to protect it.

The new design is to simply let every splintermail user have their own
ACME account.

The new server/acme/DESIGN.md reflects the new design we have, where the
splintermail server is now only responsible for managing the dns entries
for passing ACME challenges.