fix bug in dstr_recode
The ignore_partial parameter to do_dstr_recode was so confusing I picked
it wrong in dstr_recode().  Was it ignoring partials and copying them?
Or ignoring them and leaving them?

It is now called "force_end" like the base64 decoder's parameter, which
clearly indicates when it should be true and when it should not be true.

This bug could have affected:
- imap expression string decoding, if a string ended in a \" or \\
- new-key detection logic for folders whose names ended in a \n or \\