libimap: starttls and cmd/resp readers
Separating cmd and response readers allows for stronger type-checking.
This has been possible since the transition to gen.py for imap parsing,
but it didn't make sense before since the imap_engine_t was not worth
re-writing.

With the coming libengine rewrite, new imap readers will have stronger
typing.

Additionally, the libengine rewrite will make it possible to support
the STARTTLS command, which carries with it the possibility that a TLS
handshake might otherwise appear in the command stream.  Therefore, any
object which might read a STARTTLS command needs to be able to recover
all input previously fed to the command reader (the beginning of the
handshake) and the command reader must not process those bytes following
the handshake.