duvtls: add test_tls_server and support STARTTLS
STARTTLS requires that when you wrap the server, you might have to
inject preinput into the SSL object, since if the client were to
pipeline the STARTTLS command then the TLS handshake the server may have
already already read the client handshake off the wire when it reads the
STARTTLS response.

Admittedly, that would be difficult because they'd have to intercept the
OK response before passing the handshake response into their TLS
library.  But it would save an unnecessary synchronization point so
clients could be motivated to do it anyway.