libduv: make stream_i failures atomic
Before I debated sending await_cb before responding with errored
read_cbs and write_cbs, or perhaps sending the error with the first
failure, whether that be read_cb, write_cb, or await_cb.  The await_cb
pattern chosen had the benefit that there was a clear point of
"complete", which carried the error information, but passing read_cbs
and write_cbs with ok=false leaked stream state into the stream
consumer.  It was tedious for the stream consumer to keep track of the
intermediate "this stream is about to fail and I have to make sure not
to use it anymore" information.

Instead, remove the status from read_cbs and write_cbs, and pass
unfinished read and writes requests with the await_cb.  This way, there
are no intermediate failure states, and there is no dispersed error
checking (in read_cb, in write_cb, and in await_cb).  All read_cbs and
write_cbs represent successful reads and writes.

Additionally, this change removes the .readable and .writable methods of
the stream interface, which turned out to be basically useless.