libduv: improve scheduling logic
Let schedule() calls default to no-op when already scheduled, and in the
rare cases where somebody wants to explicitly

This allows a stream to effectively place itself after all other
scheduled streams, without those streams accidentally moving themselves
to the end of the line during some seemingly harmless re-schedule().

This is intended to let the libcitm-rewritten imap server and client
send write_cb's, place themselves at the end of the queue, and
potentially buffer multiple commands from their writer without the
writer having to think about it.

This still isn't bulletproof though, because if the write_cb schedules
the writer, then the writer wakes up and makes a callback to another,
higher layer, then that higher layer will schedule itself after the
imap layer, and the while delay-once-to-check-for-new-writes breaks
down.

Currently the most important case of this is passthru_t, and it actually
has a handle of the duv_scheduler_t to trump all other stream objects,
but no solution exists with the current scheduling logic that would
fully support this behavior.

We almost need subschedulers, so the imap object could own a
subscheduler, drain only the resources associated with that stream, and
execute something after that.

But at least for now, this is probably good enough.