combined sf_pair/server/fetcher/passthru into sc_t
sc_t is only about 300 loc more than server_t was, but we dropped 1700
loc and several files in the process!

Many things are atomic which were previously spread throughout three
asynchronous objects, all of the passthru_t evaporated, and now the
physical code layout tracks much of the state that used to be in
callbacks and switch statements.