libduv: duv_connect takes dstr_t host and port
It has to make a copy of the host and port, so even though the
underlying gai function takes a cstr, accepting a cstr as input saves no
copies.  On the other hand, all cstrs can be represented as dstrs, but
not vice versa, so accepting dstr_t makes the most sense.