refactor: stop auto-awaiting things
Stick to structured concurrency.  This also means that the citm_t has a
handle on all active objects (directly or indirectly) and so we can
write to dump the whole application state and not have to worry if
background closing might fail.

(there's still one exception, that the uv_citm_conn_t's .free() will
close the tcp in the background, but it's less risk to assume that libuv
has no bugs there).

Tests haven't been fixed yet.