libimap: fix bug in scanner around YYRESTORE
The issue was that relying on cursor after YYRESTORE was called is
unsafe.

The new strategy explicitly detects if a limit was hit instead.

Additionally, YYRESTORE to IMAP_INVALID_TOKEN could cause a LOG_FATAL
assertion to fail.  The assertion protects some important assumptions,
so in those cases we simply expand the IMAP_INVALID_TOKEN to include
more than a single character.