libdstr: LOG_*() functions try to avoid malloc
There was a bug where a signal handler with a LOG_INFO could cause a
hang when compiled with asan, since LOG_INFO would malloc and free a
buffer in all cases.

This makes a few optimizations:
- LOG_*() functions process the format string once for all outputs
- LOG_*() functions don't process the format string if no outputs have a
  suitable log level for the given message
- the buffer that is build is built on the stack if it fits, and only
  if it exceeds 1024 characters does it get put on the heap

This also adds a test for sb_expand() which has always had a broken heap
codepath previously.