benchmark an alternative to FMT
FMT2 is based on all lessons learned about interfaces in C.  It's much
more flexible than FMT was, and extension fmt_i's look and feel
identical to first-class fmt_i's.

It's also 30% faster in string formatting, and 70% faster in writing to
FILE*'s.  The latter speedup is large due to the innovation of a
writer_i which allows code reuse of the format logic without the
extraneous malloc or extra file copy.

Compared to snprintf and fprintf, FMT2 is consistently about 1/3 as
fast.  But it sure is nice to work with...

The benchmark code currently lives in test/test_fmt.c, but now that the
benchmark is done, FMT will be deleted, FMT2 will become FMT, and all
code will be updated, and the benchmark code will be deleted.