libacme: revamp and almost finish it
The main detail which still remains is to create and submit CSRs, but I
want to get the DNS infrastructure up and running first.

For now, the rest of the request types are complete, and there is a new
acme_t that serves a few purposes:

- it collects useful objects required for every acme call
- it auto-fetches the directory as needed
- it auto-fetches the nonce as needed, and auto-retries with a fresh
  nonce after badNonce failures
- it tracks nonces in Replay-Nonce headers automatically
- it auto-pages through lists of orders

Additionally, libacme now uses jdump() instead of FMT to compose json.