libacme: acme_close() takes a void *cb_data
acme_t is a void* because it's a very large object that is probably
better malloc'd than put on the stack.

Making it an opaque object allows hiding an enormous amount of detail
inside the compiled reqs.o.

But if that's the case, then having acme_close_cb() return a pointer to
the acme_t that got closed is just dumb.

Keep the opaque pointer, but return a void *cb_data instead of the
acme_t.