introducing libcli
libcli combines the features previously in:
- ui.c
- console_input.c
- print_help.h

Also, do_main() takes an explicit ui_i interface, rather than the old
linking tricks to replace certain files.  This also means that test_ui
compiles and links against all the modules that libcli depends on, which
is theoretically worse but also a lot simpler in practice.