libcitm: add status server
The acme_manager_t is a complex background process, which needs to
observability in order to be useful or debuggable.

The first idea was to write status files and watch them from a cli
command, but that broke down for two reasons:
  - status files would be written into a directory requiring sudo
  - status files can't easily communicate if the server is alive

In particular, the second problem clearly indicated that a
connection-oriented status server is the right approach; you
automatically find out if the server is alive, and are automatically
notified when it dies.  No point in trying to build some pid watching
technology to replicate that.

The second idea was to have a webui.  Unfortunately the webserver idea
wouldn't work with splintermail.com because splintermail.com is an
HTTPS-only domain, and I'm not very interested in a non-https domain.
But of course, serving an HTTPS domain only makes sense when the thing
is working, which defeats the purpose entirely.  Also, one of the great
benefits of the webui is that users don't need to download software, but
of course they already need the service installed so having a native
client installed is trivial.

Also, no tunneling technology could be used to solve the HTTPS problem
because a tunnel wouldn't work until after the service was configured,
and even then you still wouldn't be able to clearly tell the difference
between "misconfigured" and "broken".

So a uv_pipe_t server it is.  It can grant read permissions to anyone on
the local machine, since it doesn't expose anything sensitive.
Configuring the acme_manager_t still requires sudo, which is
appropriate.