libhttp: make duv_http_close() behave nicely
The new pattern to make writing advance_state() easy is for closers to
be idempotent, and to expose their internal state.

Specifcally, expose whether or not thing_close() will emit a close
callback at some point in the future or not.

This allows advance_state functions to not need to keep track on what
was initialized, and close everything in a sane way as part of the
cleanup sequence.