libimap: bug in write check of imap_write
imap_write has a check to cause it to fail if nothing is written.

However, its API suggests that in cases where nothing fits, it would
set *want and be just fine.

The new stream-based CITM seems to depend on that behavior, and to avoid
it would involve additional, awkward length checks before or after every
imap_write.  So just loosen up the logic requiring writing to happen; it
wouldn't make sense to enforce it if nothing would fit.