support AUTH=PLAIN and AUTH=LOGIN
AUTH=PLAIN is an imap extension required by IMAP4rev1.

It turns out we've been advertising AUTH=PLAIN but never supported it.
Also it turns out that the best way to support it is to delete the
IMAP_CMD_AUTH enum type and emit IMAP_CMD_LOGIN instead.

AUTH=LOGIN is not very hard after AUTH=PLAIN, and is a more sane
protocol anyway.

Also, this change removes the SLASH symbol from the imap parser, which
should just be normal ATOM text.  It was being allowed inside of atoms
but not at the start of atoms.