dns: CAA fixes
Define CAA for *.user.splintermail.com.  Might as well since it is the
first thing that letsencrypt checks.

Also set a 60-second TTL, in case it is broken again in the future.

Also fix the critical bit.

Also fix the CAA issuer to be letsencrypt.org rather than .com, doh.